fixed both groupId & artifactId identification renamed jcache/cache-basics to jcache/create-cache
