+++ cdi/pom.xml
@@ -10,8 +10,7 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <groupId>org.javaee8.cdi</groupId>
-    <artifactId>cdi-samples</artifactId>
+    <artifactId>cdi</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>pom</packaging>
     <name>JSR 365: CDI 2.0 Samples</name>
