+++ jsonb/pom.xml
@@ -10,8 +10,7 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <groupId>org.javaee8.json-b</groupId>
-    <artifactId>json-b-samples</artifactId>
+    <artifactId>jsonb</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>pom</packaging>
     <name>JSR 367: JSON-B Samples</name>
